refactor import data page to use functions from the import helper
