+++ alphastats/gui/utils/analysis_helper.py
@@ -425,8 +425,8 @@ def load_options():
         interpretation_options,
     )
 
-    st.session_state["plotting_options"] = plotting_options
-    st.session_state["statistic_options"] = statistic_options
+    st.session_state["plotting_options"] = plotting_options(st.session_state)
+    st.session_state["statistic_options"] = statistic_options(st.session_state)
     st.session_state["interpretation_options"] = interpretation_options
